chore(graph): extract graph-common
