--- src/CustomGraphComponent/index.ts
@@ -1,3 +1,0 @@
-export { default, Props } from './CustomGraphComponent'
-export { default as Edge, Props as EdgeProps } from './Edge'
-export { NodeProps } from './types'
