+++ src/customGraph/constants.ts
