+++ src/customGraph/index.ts
@@ -1,1 +1,3 @@
-export { default, Props } from './Graph'
+export { default, Props } from './CustomGraph'
+export { default as Edge, Props as EdgeProps } from './Edge'
+export { NodeProps } from './types'
