+++ src/graph-common/color-scheme.ts
