+++ src/graph-common/components/DrawerTabs.tsx
@@ -6,7 +6,7 @@ import { DrawerNodeLabel } from '../styles'
 import { NoData } from '../../profiler/styles'
 import { RawAggSumData } from '../../aggSum/model'
 import { JoinRates, AggSum, Profiler, Quality } from '../..'
-import { useGraphStore, setDrawerTab, closeDrawer, useGraphDispatch } from '../graph-store'
+import { useGraphStore, setDrawerTab, closeDrawer, useGraphDispatch } from '../store'
 
 const { TabPane } = Tabs
