+++ src/graph-common/components/SidebarSearch.tsx
