+++ src/graph-common/constants.ts
