+++ src/graph-common/model.ts
