+++ src/graph-common/store/index.ts
@@ -1,3 +1,4 @@
 export * from './handlers'
 export * from './model'
 export * from './StoreProvider'
+export { defaultHighlights } from './store'
