+++ src/graph-common/store/model.ts
