+++ src/graph-common/styles.ts
