+++ src/graph-common/traversal.ts
