+++ src/graph-common/usePreventDoubleClick.ts
